+++ tests/unit/Plugin/PluginAwareTraitTest.php
@@ -2,6 +2,7 @@
 
 namespace TheFrosty\PhpUnit\WpUtilities;
 
+use TheFrosty\WpUtilities\Plugin\Init;
 use TheFrosty\WpUtilities\Plugin\Plugin;
 use TheFrosty\WpUtilities\Plugin\PluginAwareTrait;
 use TheFrosty\PhpUnit\WpUtilities\Framework\TestCase;
@@ -31,7 +32,7 @@ class PluginAwareTraitTest extends TestCase
             return;
         }
 
-        $plugin = new Plugin();
+        $plugin = new Plugin(new Init());
         /** PluginAwareTrait @var PluginAwareTrait $provider */
         $provider->setPlugin($plugin);