+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2018-04-04
+### Updated
+- Changed `Init` from an abstract class to final class.
+- Removed extension of `Init` from `AbstractPlugin` and added as a DI.
+- Updated `PluginInterface` and `PluginFactory` to reflect the new changes.
+
 ## [1.0.0] - 2018-03-30
 ## Breaking Update
 ### Added